Update gluster-block.c
current coverity does not handle gcc-7's new built-in types (_Float128 specifically. No, we don't use it, but it's in header files that are included.)
To work around this I add a typedef to the generated config.h before running coverity. But for this to work, the #include "config.h" needs to be the first #include.